Interface'ify the knowledgebase model.
What is being written now is a simply in-memory-only form, but later I expect to write a disk-backed one using something simple/portable like boltdb, and perhaps later heavier types of relational database as well.
