--- model/cassandra/watch.go
@@ -1,36 +1,0 @@
-package cassandra
-
-import (
-	"polydawn.net/repeatr/model/catalog"
-)
-
-// ObserveAll: good for work planners
-// ObserveOne: convenient for catalog publishers (presumably looking at the set of stage3/runrecords under a particular stage2)
-// ObserveSet: no such interface.  harder to update asynchronously; shut up and take all and filter it yourself.
-
-/*
-	Subscribe to updates to catalogs.  Every time there's a new edition of
-	a catalog available, that catalog ID will be sent to the channel.
-
-	This subscription only makes it about as far as your local post
-	office -- you'll get notifications when a new edition makes it
-	*there*; it doesn't guarantee the publishes are actively
-	pushing notifications that far, or that the office is continuously
-	checking upstream.  For example, a catalog based on local dirs
-	might well scan continuously, but a git catalog probably doesn't
-	poll the remote ever 100ns (but it might get webhooks that
-	trigger a looksee, too!).
-
-	Catalog ID (as opposed to the full book) are passed because you typically
-	always want to interact with the latest version, and catalogs themselves
-	do not contain links to their history or any ordering hints; so, you
-	always need to respond to nudges by fetching then latest edition.
-	(Specifically, this is meant to be a safeguard if you're listening to
-	more than one buffered source of updates, and you need to be sure to
-	behave correctly even if you heard about things in the wrong order.)
-*/
-func (kb *Base) ObserveCatalogs(ch chan<- catalog.ID) {
-	kb.mutex.Lock()
-	defer kb.mutex.Unlock()
-	kb.catalogObservers = append(kb.catalogObservers, ch)
-}
